Ditto

+++ backend/scripts/docker-development.sh
@@ -2,25 +2,29 @@
 
 set -eu
 
-if getent passwd node 2>&1 >/dev/null; then
-  userdel node
-fi
-
-if getent group node 2>&1 >/dev/null; then
-  groupdel node
-fi
-
 uid=$(stat -c %u .)
 gid=$(stat -c %g .)
-groupadd -g $gid node
-useradd -u $uid -g $gid -d /home/node -s /bin/bash node
-chown -R node:node /home/node
+
+if [ $uid = 0 ]; then
+  user=root
+else
+  if getent passwd node 2>&1 >/dev/null; then
+    userdel node
+  fi
+  if getent group node 2>&1 >/dev/null; then
+    groupdel node
+  fi
+  groupadd -g $gid node
+  useradd -u $uid -g $gid -d /home/node -s /bin/bash node
+  chown -R node:node /home/node
+  user=node
+fi
 
 if ! test -d node_modules; then
   echo "Installing node modules..."
-  su -c "npm install --silent" node
+  su -c "npm install --silent" $user
 fi
 
-su -c "node scripts/seed.js" node
+su -c "node scripts/seed.js" $user
 
-su -c "npm run develop" node
+su -c "npm run develop" $user
